* [conf]: Fix xcm udev rule
The current udev rule for xcm does not work, for udev cannot find xcmddc
executable. It seems that specifying an absolute path to xcmddc in the last
rule resolves the problem.
Blindly setting group and mode for all /dev/ i2c entries is also wrong,
because the only entries to be touched are those, on which xcmddc --i2c
returns zero. Otherwise they can be for example temperature sensors etc.
There's also the security impact. The user of 'video' group cannot write to
i2c entries other than marked by xcmddc. So the group and mode assigning is
moved also to the last rule.